+++ src/main/java/subway/controller/StationController.java
@@ -1,10 +1,10 @@
-package subway.ui;
+package subway.controller;
 
 import org.springframework.http.ResponseEntity;
 import org.springframework.web.bind.annotation.*;
-import subway.application.StationService;
-import subway.dto.StationRequest;
-import subway.dto.StationResponse;
+import subway.service.StationService;
+import subway.service.dto.StationRequest;
+import subway.service.dto.StationResponse;
 
 import javax.validation.Valid;
 import java.net.URI;
