+++ src/main/java/subway/service/dto/LineResponse.java
@@ -1,4 +1,4 @@
-package subway.dto;
+package subway.service.dto;
 
 import subway.domain.Line;
 import subway.domain.Section;
