+++ src/main/java/subway/service/dto/SectionRequest.java
@@ -1,4 +1,4 @@
-package subway.dto;
+package subway.service.dto;
 
 import javax.validation.constraints.NotBlank;
 import javax.validation.constraints.NotNull;
