+++ src/main/java/subway/service/dto/StationRequest.java
@@ -1,17 +1,18 @@
-package subway.dto;
+package subway.service.dto;
+
 
 import javax.validation.constraints.NotBlank;
 
-public class LineRequest {
+public class StationRequest {
 
     @NotBlank
     private final String name;
 
-    public LineRequest(final String name) {
+    public StationRequest(final String name) {
         this.name = name;
     }
 
-    private LineRequest() {
+    private StationRequest() {
         this(null);
     }
